Added SpatialGrid.cs for fast culling extended tile layer with culling added alternative implementation of EntityLayer.cs to use spatial grid  - BUG: When using SpGr entities are aligned wrong
Reduced Draw calls from >4000 to approx 350 in demo scene

+++ WarTornLandsRefurbished/WarTornLandsRefurbished/World/Area.cs
@@ -90,7 +90,7 @@ namespace WarTornLands.World
             return result;
         }
 
-        public List<Entity> GetEntitiesAt(Vector2 position, float radius)
+        public List<Entity> GetEntitiesAt(Vector2 position, int radius)
         {
             List<Entity> result = new List<Entity>();
